working towards first immutable implementation; not finished
Based on the talk with kelly, I started to implement full tree cloning
on modification. It may be fast enough based on what little needs to
be copied on clone.

Nodes need to have clone equality without any attributes having been
set, so a per-instance ID is going to be implemented. Each tree will
maintain the next id to use on any child node, which node addition
methods will need to update. This may be a big complicated change, so
I'm committing my current work so as not to lose it.

The API isn't solid yet, and lots still needs to be implemented and
tested.