fix inflate.byAdjacencyList; implement node.parent
byAdjacencyList was broken due to the way children are processed. The
method needed to emit an array of children *as independent trees*, not
as a non-tree object to be parsed.

node.parent was pretty trivial. parents need to be assigned in two
places. It would be nice if it were just one. Cloning is a bit of a
kludge that way right now, though.