fix node children mutability
The change required older tests to change, since jasmine's `toEqual`
method doesn't seem to ever equate frozen objects with non-frozen
objects.

Also, I expected that pushing something onto a frozen array in strict
mode would throw an error, but I didn't get it to. I wonder if it's
because I'm running tests in a node environment instead of a browser.