implement `_tree.fromNode` with stub tests
shaky design decision: trees made with `_tree.fromNode` are **not**
clone-equal to the node's previous tree. Ancestry breaks with this
method. This also means that nodes are not clone-equal. I can imagine
wanting to maintain the relationships, but if implemented as the
default behavior, partial trees (fromNode) will be considered clones
of full trees, which seems like it's assuming too much about specific
usage scenarios.