implement Tree.moveNode; test it trivially
There may be some redundancy in setting parents on nodes, which is now
done in Node.clone and in the tree finalization function.

Tree.clone now steals the __nextNodeId from the original. I'm not sure
why I would've wanted it this way, but no tests have failed on
it. Maybe it was intentional? Anyhow, no tests fail now that it is
this way, so either it's fine, redundant. or untested. Probably
untested.