simplify example; better+redo benchmarks; add jsfiddle link
The logging made the example kindof overwhelming, so I axed it.

The inflation child-count example made me wonder what the performance
was like for increasing child nodes. Up to 16, it looks
logarithmic. It probably is, since 1024-wide nodes inflate at ~10/sec.

The fiddle link is a flot chart. It really shows the trend.