reimplement `Node.extend` on Backbone's internal extend; remove cruft
The cruft was lingering references to `Node.clone`, which needed to
instead reference the current (dynamic) node subclass (if any).

Backbone's `extend` implementation fit very well for my Node subclass
needs, and it needed little modification. I am considering replacing
the `__super__` reference with a reference to the parent constructor
instead of the contructor prototype. Seems more useful. Not urgent.