implement `_tree.Tree.extend`; refactor `new Tree` to accomodate
`Tree` can now be extended in the same way as `Node`. I'm not at my
best right now, and think that I might not have implemented it
correctly in all cases. `_tree.fromNode` was a picky instance, where
the new tree could reasonably be an instance of the node's original
tree's class, or an instance of the passed-in default tree class. I
went with the default option , since `fromNode` should not assume the
new tree will necessarily be like the previous.